NF: move CheckCardSelection where its used
